day 1

+++ Introductory Problems/Weird Algo.cpp
@@ -4,9 +4,9 @@
 using namespace std;
 
 int main() {
-    int n;
+    long long n;
     cin >> n;
-    vector<int> arr;
+    vector<long long> arr;
     arr.push_back(n);
 
     while (n != 1) {
